rag-engine/requirements.txt
langchain
openai
faiss-cpu
python-dotenv
websockets
